refactor: split system theme service into modules
